Removing default code in App.js and updating the file
I removed the default code in App.js and updated it to show different pages that we created depending on the link. So far, since we have the home page and add products page, I only added those two links. Add more links as we complete more features of the website.