Displaying the products uploaded by users in homepage
I've added the code necessary to display the products uploaded by users on our homepage, so then users can see what's available on the market. The actual part of the code that takes each item and displays their image, name and price are in the Products.js file. I might move the entire section where I get each item in the database and store it in an array, into the Products.js folder to make this file clean as possible.