Adding ProductInfo.js to the repo
This file contains the implementation for opening product info pages. You can just click on the product and it'll automatically redirect you to it. The product info page displays the seller of the product, but this will only work for newly uploaded products with my implementation because I added a new field for products that has the seller's name and uid.