Adding signout feature and fixing viewing profiles
Created a signout feature as a button within the navbar so users can explicitly signout after using the website. The website does not yet automatically sign users out which needs to be added. I removed the navbar as well from the register and log in page as it doesn't make sense for users to have access to it when they're not logged in yet. Viewing other profiles was creating an error because of useEffect not triggering before accessing profiles within the database, and this was fixed as well.