Adding Category Dropdown menu
I added a category dropdown menu in our navbar, as well as remove the login button, since it doesn't make sense to login after already logging in. To implement the dropdown menu, I had to make a major change to Navbar, where I map a list of navitems on to the navbar, and return that. Check the new files I added to see how I implemented it if confused. Styling has not been finished, so it doesn't look great, but the functionality should work.