Adding the delete product feature
This commit adds the feature of being able to delete your own products so that once you sold an item, you can remove them. Because we use the Products file to also display products in the home page and categories, I needed to create a new file (ProfileProducts) to implement this. There should be a button under your products in profile that says delete item now.