Deleting chats and images in storage after deleting item
This commit adds on to the previous commit of deleting products, where now it properly deletes any chats related with the deleted product, as well as the image within the firebase storage. I needed to add another field when adding products in order to refer to the image path within storage, so I will delete all the old items in order for this feature to work.